Refactor KLIFS API/processor
