+++ indra/sources/klifs/__init__.py
@@ -1,29 +1,9 @@
-# indra/sources/klifs/__init__.py
-
 """INDRA source integration for KLIFS (Kinase–Ligand Interaction Fingerprints
 and Structures).
 
 Public API is primarily a thin, Swagger-mirroring layer defined in .api.
 """
 
-from .api import (
-    get_kinase_groups,
-    get_kinase_families,
-    get_kinase_names,
-    get_kinase_information,
-    get_kinase_id,
-    get_ligands_list,
-    get_bioactivities_for_ligand,
-    process_bioactivities_for_ligand,
-)
+from .api import process_ligand
 
-__all__ = [
-    "get_kinase_groups",
-    "get_kinase_families",
-    "get_kinase_names",
-    "get_kinase_information",
-    "get_kinase_id",
-    "get_ligands_list",
-    "get_bioactivities_for_ligand",
-    "process_bioactivities_for_ligand",
-]
+__all__ = ["process_ligand"]
