+++ indra/sources/klifs/client.py
@@ -1,7 +1,3 @@
-# indra/sources/klifs/client.py
-
-from __future__ import annotations
-
 import logging
 from dataclasses import dataclass
 from typing import Any, Dict, Iterable, List, Optional, Union
