+++ indra/sources/klifs/processor.py
@@ -1,13 +1,9 @@
-# indra/sources/klifs/processor.py
-
-from __future__ import annotations
-
-from dataclasses import dataclass, field
-from typing import Any, Dict, List, Optional
-
-from indra.statements import Agent, Complex, Evidence, Inhibition, Statement
+from typing import Any, Dict, List, Optional, Union, Iterable
 
 import html
+
+from indra.statements import Agent, Complex, Evidence, Inhibition, Statement
+from .client import KlifsClient
 
 
 def _safe_str(x: Any) -> Optional[str]:
@@ -25,7 +21,9 @@ def _norm_measurement_type(x: Any) -> Optional[str]:
     return s.upper().replace(" ", "")
 
 
-@dataclass
+_default_client = KlifsClient()
+
+
 class KlifsProcessor:
     """Processor for turning KLIFS BioactivityDetails into INDRA Statements.
 
@@ -47,12 +45,19 @@ class KlifsProcessor:
     All quantitative fields are preserved in Evidence annotations.
     """
 
-    bioactivities: List[Dict[str, Any]]
-    ligand_id: Optional[int] = None
-    ligand_pdb: Optional[str] = None
-    ligand_details: Optional[Dict[str, Any]] = None
 
-    statements: List[Statement] = field(default_factory=list)
+    def __init__(self, ligand_id, ligand_pdb, ligand_details):
+        self.ligand_id = ligand_id
+        self.ligand_pdb = ligand_pdb
+        self.ligand_details = ligand_details
+
+        self. bioactivities = bioactivities = get_bioactivities_for_ligand(
+            ligand_id=ligand_id,
+            ligand_pdb=ligand_pdb,
+            client=_default_client,
+        )
+
+        self.statements = []
 
     def extract_statements(self) -> None:
         """Extract INDRA Statements into `self.statements`."""
@@ -154,3 +159,106 @@ class KlifsProcessor:
                 ann[k] = v
 
         return ann
+
+
+def get_kinase_families(
+        kinase_group: Optional[Union[str, Iterable[str]]] = None,
+        client: KlifsClient = _default_client,
+) -> List[str]:
+    """Get the list of kinase families from KLIFS.
+
+    Swagger mapping: GET /kinase_families
+    """
+    return client.kinase_families(kinase_group=kinase_group)
+
+
+def get_kinase_names(
+        kinase_group: Optional[Union[str, Iterable[str]]] = None,
+        kinase_family: Optional[Union[str, Iterable[str]]] = None,
+        species: Optional[str] = None,
+        client: KlifsClient = _default_client,
+) -> List[Dict[str, Any]]:
+    """Get kinase names (HGNC gene symbols) and associated IDs from KLIFS.
+
+    Swagger mapping: GET /kinase_names
+    """
+    return client.kinase_names(
+        kinase_group=kinase_group,
+        kinase_family=kinase_family,
+        species=species,
+    )
+
+
+def get_kinase_information(
+        kinase_ids: Optional[Union[int, Iterable[int]]] = None,
+        species: Optional[str] = None,
+        client: KlifsClient = _default_client,
+) -> List[Dict[str, Any]]:
+    """Get kinase information records from KLIFS.
+
+    Swagger mapping: GET /kinase_information
+    """
+    return client.kinase_information(kinase_id=kinase_ids, species=species)
+
+
+def get_kinase_id(
+        kinase_name: Union[str, Iterable[str]],
+        species: Optional[str] = None,
+        client: KlifsClient = _default_client,
+) -> List[int]:
+    """Resolve one or more kinase names to KLIFS kinase_ID integers.
+
+    Swagger mapping: GET /kinase_ID
+
+    Notes
+    -----
+    The Swagger endpoint returns KinaseInformation-like records; this helper
+    extracts and returns only the `kinase_ID` integer(s).
+    """
+    rows = client.kinase_id(kinase_name=kinase_name, species=species)
+    out: List[int] = []
+    for row in rows or []:
+        kid = row.get("kinase_ID")
+        if kid is not None:
+            out.append(int(kid))
+    return out
+
+
+def get_ligands_list(
+        kinase_id: Optional[Union[int, Iterable[int]]] = None,
+        client: KlifsClient = _default_client,
+) -> List[Dict[str, Any]]:
+    """Get ligandDetails records from KLIFS.
+
+    Swagger mapping: GET /ligands_list
+    """
+    return client.ligands_list(kinase_id=kinase_id)
+
+
+def get_bioactivities_for_ligand(
+        ligand_id: Optional[int] = None,
+        ligand_pdb: Optional[str] = None,
+        client: KlifsClient = _default_client,
+) -> List[Dict[str, Any]]:
+    """Get bioactivity records for a ligand from KLIFS.
+
+    KLIFS exposes two separate Swagger endpoints that return the same *kind* of
+    record (BioactivityDetails) for a ligand, differing only by identifier:
+
+    - GET /bioactivity_list_id   (use when `ligand_id` is provided)
+    - GET /bioactivity_list_pdb  (use when `ligand_pdb` is provided)
+
+    This convenience function dispatches to the correct endpoint.
+
+    Raises
+    ------
+    ValueError
+        If neither `ligand_id` nor `ligand_pdb` is provided.
+    """
+    if ligand_id is not None:
+        return client.bioactivity_list_id(ligand_id)
+    if ligand_pdb is not None:
+        return client.bioactivity_list_pdb(ligand_pdb)
+    raise ValueError("Provide ligand_id or ligand_pdb.")
+
+
